requirements.txt
cmake==3.13.2
dj-database-url==0.5.0
Django==1.11.17
django-admin==1.3.2
django-cors-headers==2.4.0
django-excel-response2==2.0.8
django-six==1.0.4
djangorestframework==3.9.0
gunicorn==19.9.0
psycopg2==2.7.6.1
pytz==2018.7
screen==1.0.1
whitenoise==4.1.2
xlwt==1.3.0
